refactor(ir): ✓ lift the pure-AST loop-shape closure below the IR (#4601 route 1)
`src/codegen/statements/loop-analysis.ts` is pure AST from top to bottom — no
CodegenContext, no checker, no emission — yet five `src/ir/` modules imported it
across the layer boundary: from-ast, char-read-loop, i32-pure-bitwise,
fixed-literal-loop-proof and analysis/i32-slots. #4601's progress log recorded
why it could not simply move: its own two dependencies,
`collectReferencedIdentifiers` (closures.ts, 3,984 LOC) and
`collectPatternBindingNames` (statements/tdz.ts), lived in CodegenContext-typed
homes, so no relocation-only move set terminated in a leaf.

This lifts that leaf closure. New `src/ir/analysis/ast-scope.ts` holds the pure
scope / free-variable walks verbatim (isFunctionScopeBoundary,
collectBindingPatternNames, collectPatternBindingNames, collectFunctionOwnLocals,
collectVarAndTopLevelDecls, collectReferencedIdentifiers); the memoizing oracle
`binding-info.ts` moves with them; `loop-analysis.ts` becomes
`src/ir/analysis/loop-shape.ts`. Its dependency set is now `ts-api` + `ast-scope`
— a genuine leaf.

Old homes re-export every moved name, so no codegen consumer's import changes
meaning. Pure motion: `prove-emit-identity check` reports IDENTICAL across all
60 (file,target) pairs, ts7 typecheck clean.

check:ir-layering 92 -> 87 import lines, 19 -> 17 files; no new file with
codegen imports.


✓
