+++ src/ir/analysis/i32-slots.ts
@@ -64,7 +64,7 @@
 import { forEachChild, ts } from "../../ts-api.js";
 import type { IrBinop } from "../nodes.js";
 import { collectI32CoercedLocals } from "../../codegen/analysis/i32-coerced-locals.js";
-import { detectI32LoopVar } from "../../codegen/statements/loop-analysis.js";
+import { detectI32LoopVar } from "./loop-shape.js";
 
 const I32_MIN = -2147483648;
 const I32_MAX = 2147483647;
