+++ src/ir/from-ast.ts
@@ -59,7 +59,7 @@ import { evaluateConstantCondition } from "../codegen/statements/control-flow.js
 import type { IrClassInstanceInitializer } from "./class-instance-initializers.js";
 // #2766 — reuse the legacy counted-loop proof predicates (pure AST analysis, no
 // codegen state) to port the `safeIndexedArrays` in-bounds proof into the IR.
-import { isIncreasingStep, loopBodyMutatesIndexOrArray } from "../codegen/statements/loop-analysis.js";
+import { isIncreasingStep, loopBodyMutatesIndexOrArray } from "./analysis/loop-shape.js";
 // (#3741) native-i32 slot storage for provably-int32 mutable locals
 import {
   COMPOUND_TO_BITWISE_TOKEN,
